scripts/requirements.txt
msgpack
msgpack-numpy
pyserial